@@ -81,11 +81,11 @@ class IndicatorsController < ApplicationController
         i = find_instance
         i.value = params[:indicator][:value]
         i.value_will_change!
-        attributes = :value
+        attributes = {:value => i.value}
         if params[:indicator][:last_update]
           i.last_update = params[:indicator][:last_update]
           i.last_update_will_change!
-          attributes = [:value, :last_update]
+          attributes = {:value => i.value, :last_update => i.last_update}
         end
         hobo_update(i, :attributes => attributes) do
           redirect_to this.objective.area.hoshin if valid? && !request.xhr?
