@@ -1,6 +1,8 @@
 Hoshinplan::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
 
+  Rails.application.routes.default_url_options[:host] = 'www.hoshinplan.com'
+
   # Code is not reloaded between requests
   config.cache_classes = true
 
@@ -73,7 +75,7 @@ Hoshinplan::Application.configure do
   # config.threadsafe!
 
   # Enable locale fallbacks for I18n (makes lookups for any locale fall back to
-  # the I18n.default_locale when a translation can not be found)
+  # the I18n.default_locale between a translation can not be found)
   config.i18n.fallbacks = true
 
   # Send deprecation notices to registered listeners
